[#1] feat: 카테고리 목록 조회 기능 및 테스트 추가
