+++ src/main/java/com/oheat/shop/repository/CategoryRepository.java
@@ -1,6 +1,7 @@
 package com.oheat.shop.repository;
 
 import com.oheat.shop.entity.CategoryJpaEntity;
+import java.util.List;
 import java.util.Optional;
 
 public interface CategoryRepository {
@@ -8,4 +9,6 @@ public interface CategoryRepository {
     void save(CategoryJpaEntity categoryJpaEntity);
 
     Optional<CategoryJpaEntity> findByName(String name);
+
+    List<CategoryJpaEntity> findAll();
 }
