+++ src/main/java/com/oheat/shop/repository/CategoryRepositoryImpl.java
@@ -1,6 +1,7 @@
 package com.oheat.shop.repository;
 
 import com.oheat.shop.entity.CategoryJpaEntity;
+import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
@@ -19,5 +20,10 @@ public class CategoryRepositoryImpl implements CategoryRepository {
     @Override
     public Optional<CategoryJpaEntity> findByName(String name) {
         return categoryJpaRepository.findByName(name);
+    }
+
+    @Override
+    public List<CategoryJpaEntity> findAll() {
+        return categoryJpaRepository.findAll();
     }
 }
