+++ src/test/java/com/oheat/shop/MemoryCategoryRepository.java
@@ -3,6 +3,7 @@ package com.oheat.shop;
 import com.oheat.shop.entity.CategoryJpaEntity;
 import com.oheat.shop.repository.CategoryRepository;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
@@ -18,5 +19,10 @@ public class MemoryCategoryRepository implements CategoryRepository {
     @Override
     public Optional<CategoryJpaEntity> findByName(String name) {
         return Optional.ofNullable(categories.get(name));
+    }
+
+    @Override
+    public List<CategoryJpaEntity> findAll() {
+        return categories.values().stream().toList();
     }
 }
